# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1.0/products/feirante/{id}": {
      "get": {
        "tags": [
          "Produto"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Produto"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "img": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "unit": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "finalPrice": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "marketVendorId": {
                  "example": "any"
                },
                "stand": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1.0/marketcondominium/feirante/{id}": {
      "get": {
        "tags": [
          "FeiraCondominio"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}